refactored yuma config into main yuma file
