+++ puma/hlplots/__init__.py
@@ -3,16 +3,15 @@
 from __future__ import annotations
 
 from puma.hlplots.aux_results import AuxResults
-from puma.hlplots.configs import PlotConfig
 from puma.hlplots.results import Results
 from puma.hlplots.tagger import Tagger
 from puma.hlplots.yutils import combine_suffixes, get_included_taggers
 
 __all__ = [
     "AuxResults",
-    "PlotConfig",
     "Results",
     "Tagger",
+    "YumaConfig",
     "combine_suffixes",
     "get_included_taggers",
 ]
