+++ puma/hlplots/yuma.py
@@ -1,14 +1,22 @@
+
 from __future__ import annotations
 
 import argparse
 import os
 from pathlib import Path
+from dataclasses import dataclass, field
+from datetime import datetime
 
 import yaml
 from yamlinclude import YamlIncludeConstructor
 
-from puma.hlplots import PlotConfig, combine_suffixes, get_included_taggers
+from puma.hlplots import  combine_suffixes, get_included_taggers, Results, Tagger
+from puma.hlplots.yutils import get_tagger_name
 from puma.utils import logger
+
+
+
+
 
 ALL_PLOTS = ["roc", "scan", "disc", "probs", "peff"]
 
@@ -36,15 +44,97 @@ def get_args(args):
     return parser.parse_args(args)
 
 
-def make_plots(plots, plt_cfg):
-    for plot in plt_cfg.plots:
-        if not (plot["plot_type"] in plots and plot["signal"] == plt_cfg.signal):
-            continue
-        plt_cfg.results.taggers, all_taggers, inc_str = get_included_taggers(plt_cfg.results, plot)
-        plot_kwargs = plot.get("plot_kwargs", {})
-        plot_kwargs["suffix"] = combine_suffixes([plot_kwargs.get("suffix", ""), inc_str])
-        plt_cfg.results.make_plot(plot["plot_type"], plot_kwargs)
-        plt_cfg.results.taggers = all_taggers
+
+@dataclass
+class YumaConfig:
+    config_path: Path
+    plot_dir: Path
+
+    results_config: dict[str, dict[str, str]]
+    taggers_config: dict
+    taggers: list[str] | list[Tagger] | None = None
+
+    timestamp: bool = False
+    base_path: Path = None
+
+    plots: list[dict[str, dict[str, str]]] = field(default_factory=list)
+
+    def __post_init__(self):
+        # Define a plot directory based on the plot config file name, and a date time
+        plot_dir_name = self.config_path.stem
+        if self.timestamp:
+            date_time_file = datetime.now().strftime("%Y%m%d_%H%M%S")
+            plot_dir_name += "_" + date_time_file
+        self.plot_dir_final = Path(self.plot_dir) / plot_dir_name
+
+        self.plots = [p["args"] for p in self.plots]
+
+        for k, kwargs in self.taggers_config.items():
+            kwargs["yaml_name"] = k
+
+    @classmethod
+    def load_config(cls, path: Path, **kwargs) -> YumaConfig:
+        if not path.exists():
+            raise FileNotFoundError(f"Config at {path} does not exist")
+        with open(path) as f:
+            config = yaml.safe_load(f)
+        return cls(config_path=path, **config, **kwargs)
+
+    def get_results(self):
+        """
+        Create the high-level 'Results' object from the config file, using the
+        previously set signal and sample. Iterates and loads all models in the config
+        file, and adds them.
+        """
+        kwargs = self.results_config
+        kwargs["signal"] = self.signal
+        kwargs["perf_vars"] = self.peff_vars
+
+        sample_path = kwargs.pop("sample_path", None)
+        if self.base_path and sample_path:
+            sample_path = self.base_path / sample_path
+
+        # Instantiate the results object
+        results = Results(**kwargs)
+
+        # Add taggers to results, then bulk load
+        for key, t in self.taggers_config.items():
+            # if the a sample is not defined for the tagger, use the default sample
+            if not sample_path and not t.get("sample_path", None):
+                raise ValueError(f"No sample path defined for tagger {key}")
+            if sample_path and not t.get("sample_path", None):
+                t["sample_path"] = sample_path
+            # Allows automatic selection of tagger name in eval files
+            t["name"] = get_tagger_name(
+                t.get("name", None), t["sample_path"], key, results.flavours
+            )
+            results.add(Tagger(**t))
+
+        results.load()
+        self.results = results
+
+    @property
+    def signals(self):
+        """Iterates all plots in the config and returns a list of all signals."""
+        return list({p["signal"] for p in self.plots})
+
+    @property
+    def peff_vars(self):
+        """Iterates plots and returns a list of all performance variables."""
+        return list({
+            p["plot_kwargs"].get("perf_var", "pt") for p in self.plots if p["plot_type"] == "peff"
+        })
+
+
+    def make_plots(self, plots):
+        for plot in self.plots:
+            if not (plot["plot_type"] in plots and plot["signal"] == self.signal):
+                continue
+            self.results.taggers, all_taggers, inc_str = get_included_taggers(self.results, plot)
+            plot_kwargs = plot.get("plot_kwargs", {})
+            plot_kwargs["suffix"] = combine_suffixes([plot_kwargs.get("suffix", ""), inc_str])
+            self.results.make_plot(plot["plot_type"], plot_kwargs)
+            self.results.taggers = all_taggers
 
 
 def main(args=None):
@@ -54,7 +144,7 @@ def main(args=None):
     YamlIncludeConstructor.add_to_loader_class(
         loader_class=yaml.SafeLoader, base_dir=config_path.parent
     )
-    plt_cfg = PlotConfig.load_config(config_path, base_path=args.dir)
+    yuma = YumaConfig.load_config(config_path, base_path=args.dir)
 
     # select and check plots
     plots = args.plots if args.plots else ALL_PLOTS
@@ -62,22 +152,22 @@ def main(args=None):
         raise ValueError(f"Unknown plot types {missing}, choose from {ALL_PLOTS}")
 
     # select and check signals
-    signals = args.signals if args.signals else plt_cfg.signals
-    if missing := [s for s in signals if s not in plt_cfg.signals]:
-        raise ValueError(f"Unknown signals {missing}, choose from {plt_cfg.signals}")
+    signals = args.signals if args.signals else yuma.signals
+    if missing := [s for s in signals if s not in yuma.signals]:
+        raise ValueError(f"Unknown signals {missing}, choose from {yuma.signals}")
 
-    logger.info(f"Plotting in {plt_cfg.plot_dir_final}")
+    logger.info(f"Plotting in {yuma.plot_dir_final}")
 
     logger.info("Instantiating Results")
-    plt_cfg.signal = signals[0]
-    plt_cfg.get_results()  # only run once
+    yuma.signal = signals[0]
+    yuma.get_results()  # only run once
     for signal in signals:
         logger.info(f"Plotting signal {signal}")
-        plt_cfg.signal = signal
-        plt_cfg.results.set_signal(signal)
-        plt_cfg.results.output_dir = plt_cfg.plot_dir_final / f"{signal[0]}tagging"
-        os.makedirs(plt_cfg.results.output_dir, exist_ok=True)
-        make_plots(plots, plt_cfg)
+        yuma.signal = signal
+        yuma.results.set_signal(signal)
+        yuma.results.output_dir = yuma.plot_dir_final / f"{signal[0]}tagging"
+        os.makedirs(yuma.results.output_dir, exist_ok=True)
+        yuma.make_plots(plots)
 
 
 if __name__ == "__main__":
